WP2.1: src/ layout + PEP8 geometry classes with deprecation shim
- Package moves to src/gds_fdtd (pyproject wheel target, CI mypy path,
  docs sys.path updated).
- core.py split: geometry primitives move to gds_fdtd.geometry with
  PascalCase names (port->Port, structure->Structure, region->Region,
  component->Component, layout->LayoutSource). technology and the legacy
  s-parameter classes remain in core.py until WP2.2/WP2.4.
- gds_fdtd.core serves the old lowercase names via a PEP 562 module
  __getattr__ that emits DeprecationWarning (removal at v1.0); helper
  functions (unchanged names) re-export silently.
- All internal imports/call sites renamed; package __init__ exports the
  new names; tests updated (test_core aliases the new classes to keep
  its 950-line body stable; monkeypatch target follows the class).
- New tests/test_deprecation.py locks the shim behavior. Goldens
  unchanged: identical geometry through the rename.
