+++ src/gds_fdtd/solver_tidy3d.py
