+++ src/gds_fdtd/sparams.py
